refactor FE little bit
