--- frontend/src/components/Form/Form.css
@@ -1,5 +1,0 @@
-.wrapper {
-  margin: 6rem 4rem;
-  color: #137dc5;
-  font-size: 16pt;
-}
