+++ frontend/src/components/Splash/Splash.js
@@ -6,7 +6,7 @@ import "./Splash.css";
 const Splash = () => (
   <div className="flex flex-col justify-center h-screen mx-24">
     <Logo className="w-full" />
-    <Link to="/form" className="flex mx-auto mt-12">
+    <Link to="/generalInfo" className="flex mx-auto mt-12">
       <button className="startJourney px-5 w-auto whitespace-nowrap">
         Start your Adventure!
       </button>
